+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ TIPP3 v0.3
 #. Changed the default placement method for TIPP3 from BSCAMPP(epa-ng) to
    BSCAMPP(pplacer), which is more accurate with similar scalability and speed.
    You can change this setting with ``--bscampp-mode {epa-ng,pplacer}``.
+#. Updated refpkg version to 1.2 (i.e., the refpkg downloaded with
+   ``run_tipp3.py download_refpkg ...``). 
 
 TIPP3 v0.3b
 -----------